+++ tests/conftest.py
@@ -4,6 +4,12 @@ import pytest
 
 
 @pytest.fixture
-def test_data_dir() -> Path:
-    """Return the absolute path to the test_data directory."""
-    return (Path(__file__).parent / 'test_data').resolve()
+def cropped_despeckled_data_dir() -> Path:
+    """Return the absolute path to the cropped despeckled data directory."""
+    return (Path(__file__).parent / 'test_data' / 'T009_019294_IW2_cropped_tv').resolve()
+
+
+@pytest.fixture
+def cropped_vh_data_dir() -> Path:
+    """Return the absolute path to the cropped data directory."""
+    return (Path(__file__).parent / 'test_data' / 'T009_019294_IW2_cropped').resolve()