+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] - 2025-02-20
+
+### Fixed
+* Pyproject.toml and environment.yml to require python 3.12+
+* Pydantic model_config to allow arbitrary types (update for v2)
+
+### Removed 
+* `mpire` dependency
+
+### Added
+* Test for despeckling of cropped data
+
+
 ## [0.0.6] - 2025-01-18
 
 * Fixed default keyword argument for `merge_categorical_arrays` from `mode` (not possible with rasterio) to `min`.